requirements.txt
streamlit~=1.39.0
google-generativeai
python-dotenv~=1.0.1
langchain~=0.3.7
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
uvicorn~=0.32.0
fastapi~=0.115.4
pydantic~=2.9.2
pdfplumber